Fix reresolve logic
The previous logic only worked on links in atom feeds that contained multiple link elements. This changes how we process those multiple elements so that we can reresolve single element links, as well.

This bug was obscured because it only affected the links in the first item in a feed.

Resolves #237